move much of the lens/microlens stuff to one module
