+++ src/Codec/Xlsx/Formatted.hs
@@ -35,14 +35,7 @@ module Codec.Xlsx.Formatted
   , condfmtStopIfTrue
   ) where
 
-#ifdef USE_MICROLENS
-import Lens.Micro
-import Lens.Micro.Mtl
-import Lens.Micro.TH
-import Lens.Micro.GHC ()
-#else
-import Control.Lens
-#endif
+import Codec.Xlsx.LensCompat
 import Control.Monad (forM, guard)
 import Control.Monad.State hiding (forM_, mapM)
 import Data.Default
