+++ src/Codec/Xlsx/Types/AutoFilter.hs
@@ -9,11 +9,7 @@ module Codec.Xlsx.Types.AutoFilter where
 
 import Control.Arrow (first)
 import Control.DeepSeq (NFData)
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens (makeLenses)
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Data.Bool (bool)
 import Data.ByteString (ByteString)
 import Data.Default
