+++ src/Codec/Xlsx/Types/ConditionalFormatting.hs
@@ -43,11 +43,7 @@ module Codec.Xlsx.Types.ConditionalFormatting
 
 import Control.Arrow (first, right)
 import Control.DeepSeq (NFData)
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens (makeLenses)
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Data.Bool (bool)
 import Data.ByteString (ByteString)
 import Data.Default
