+++ src/Codec/Xlsx/Types/Drawing.hs
@@ -13,11 +13,7 @@ module Codec.Xlsx.Types.Drawing where
 
 import Control.Arrow (first)
 import Control.DeepSeq (NFData)
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens.TH
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Data.ByteString.Lazy (ByteString)
 import Data.Default
 import qualified Data.Map as M
