+++ src/Codec/Xlsx/Types/Drawing/Chart.hs
@@ -7,11 +7,7 @@ module Codec.Xlsx.Types.Drawing.Chart where
 
 import GHC.Generics (Generic)
 
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens.TH
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Control.DeepSeq (NFData)
 import Data.Default
 import Data.Maybe (catMaybes, listToMaybe, maybeToList)
