+++ src/Codec/Xlsx/Types/RichText.hs
@@ -32,11 +32,7 @@ module Codec.Xlsx.Types.RichText (
 
 import GHC.Generics (Generic)
 
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens hiding (element)
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Control.Monad
 import Control.DeepSeq (NFData)
 import Data.Default
