+++ src/Codec/Xlsx/Types/SheetViews.hs
@@ -49,11 +49,7 @@ module Codec.Xlsx.Types.SheetViews (
 
 import GHC.Generics (Generic)
 
-#ifdef USE_MICROLENS
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens (makeLenses)
-#endif
+import Codec.Xlsx.LensCompat (makeLenses)
 import Control.DeepSeq (NFData)
 import Data.Default
 import Data.Maybe (catMaybes, maybeToList, listToMaybe)
