+++ src/Codec/Xlsx/Types/StyleSheet.hs
@@ -129,12 +129,7 @@ module Codec.Xlsx.Types.StyleSheet (
   , firstUserNumFmtId
   ) where
 
-#ifdef USE_MICROLENS
-import Lens.Micro
-import Lens.Micro.TH (makeLenses)
-#else
-import Control.Lens hiding (element, elements, (.=))
-#endif
+import Codec.Xlsx.LensCompat (makeLenses, (&), (.~))
 import Control.DeepSeq (NFData)
 import Data.Default
 import Data.Map.Strict (Map)
